+++ hutool-core/src/test/java/cn/hutool/core/io/resource/ResourceUtilTest.java
@@ -1,5 +1,6 @@
 package cn.hutool.core.io.resource;
 
+import cn.hutool.core.io.IoUtil;
 import org.junit.Assert;
 import org.junit.Test;
 
@@ -9,5 +10,13 @@ public class ResourceUtilTest {
 	public void readXmlTest(){
 		final String str = ResourceUtil.readUtf8Str("test.xml");
 		Assert.assertNotNull(str);
+	}
+
+	@Test
+	public void stringResourceTest(){
+		final StringResource stringResource = new StringResource("testData", "test");
+		Assert.assertEquals("test", stringResource.getName());
+		Assert.assertArrayEquals("testData".getBytes(), stringResource.readBytes());
+		Assert.assertArrayEquals("testData".getBytes(), IoUtil.readBytes(stringResource.getStream()));
 	}
 }