+++ CHANGELOG.md
@@ -24,6 +24,8 @@
 * 【cache  】     增加CacheListener（issue#1257@Github）
 * 【core   】     TimeInterval支持分组（issue#1238@Github）
 * 【core   】     增加compile包（pr#1243@Github）
+* 【core   】     增加ResourceClassLoader、CharSequenceResource、FileObjectResource
+* 【core   】     修改IoUtil.read(Reader)逻辑默认关闭Reader
 
 ### Bug修复
 * 【cron   】     修复CronTimer可能死循环的问题（issue#1224@Github）